feat: update validation to make schemas related
The idea is to have a base schema with all fields and then cascade to a create schema and an update schema.

The create schema should accept all fields as they are, except for the ones set automatically.

The update should be similar to the create, minus fields that are not changeable, and making all fields optional